Removing temp files */

+++ userprog/syscall.c
@@ -613,7 +613,7 @@ static void system_mmap (struct intr_frame *f, int fd, void *uaddr){
 	   memory*/
 	for(i = 0, temp_ptr = (uint8_t*)uaddr; i < num_pages;
 			i ++, temp_ptr += PGSIZE){
-		ASSERT((temp_ptr % PGSIZE) == 0);
+		ASSERT(((uint32_t)temp_ptr % PGSIZE) == 0);
 		if(!pagedir_setup_demand_page(pd, temp_ptr, PTE_AVL_MMAP,
 				(uint32_t)temp_ptr, true)){
 			/* This virtual address cannot be allocated so we have an error...
